test: move sport edit, list and delete functional tests into their own controller
